requirements.txt
Django==1.11.12
django-crispy-forms==1.7.2
django-filter==1.1.0
djangorestframework==3.8.2
psycopg2==2.7.5
pytz==2018.4
